requirements.txt
pyaudio==0.2.14
openai==1.3.0
elevenlabs==0.2.26
python-dotenv==1.0.0
numpy==1.24.3
pygame==2.5.2
requests==2.31.0
